requirements.txt
xlwings >= 0.15.8
koala2 ==  0.0.31
numpy >= 1.14.5
pandas >= 0.25.0
openpyxl==2.5.3
python-harvest-2>=1.0.0
networkx == 2.1
